editor done, v1

+++ public/javascripts/recipes/edit.js
@@ -1,10 +1,11 @@
-var tabledata = [{}] ;//= [{ id: 1, ingredient: " ", isflour: false, bp: 1.5, wt: 10 }];
+loadingCursor(true);
+var tabledata = [{}]; //= [{ id: 1, ingredient: " ", isflour: false, bp: 1.5, wt: 10 }];
 var searchTable;
 var tableLength = 0;
 var defaultSort = [
-  {column: "ingredient", dir:"asc"}, 
-  {column: "wt", dir:"dec"},
-  {column: "isflour", dir:"dec"},
+  { column: "ingredient", dir: "asc" },
+  { column: "wt0", dir: "dec" },
+  { column: "isflour", dir: "dec" },
 ];
 var table = new Tabulator("#example", {
   history: true,
@@ -12,7 +13,7 @@ var table = new Tabulator("#example", {
   height: "100%", // set height of table (in CSS or here), this enables the Virtual DOM and improves render speed dramatically (can be any valid css height value)
   data: tabledata, //assign data to table
   //layout:"fitColumns", //fit columns to width of table (optional)
-  initialSort:defaultSort,
+  initialSort: defaultSort,
   columns: [
     //Define Table Columns
     {
@@ -32,15 +33,14 @@ var table = new Tabulator("#example", {
           searchTable = await fuzzySearch(term);
           searchTable.forEach((val, idx) => {
             matches.push(val.name);
-            if(idx >= 10)
-              return
+            if (idx >= 10) return;
           });
           return matches;
         },
       },
     },
-    { title: "id", field: "id", visible: true },
-    
+    { title: "id", field: "id", visible: false },
+
     {
       title: `<i class="material-icons">delete</i>`,
       field: "delete",
@@ -49,13 +49,12 @@ var table = new Tabulator("#example", {
       headerSort: false,
       maxWidth: 100,
       formatterParams: {
-        tickElement:`<i class="material-icons">delete_forever</i>`,
-        crossElement:`<i class="material-icons red-text">delete_forever</i>`,
+        tickElement: `<i class="material-icons">delete_forever</i>`,
+        crossElement: `<i class="material-icons red-text">delete_forever</i>`,
       },
       cellClick: function (e, cell) {
         let row = cell.getRow();
-        if (table.rowManager.activeRowsCount > 1)
-          row.delete();
+        if (table.rowManager.activeRowsCount > 1) row.delete();
       },
     },
     {
@@ -64,12 +63,14 @@ var table = new Tabulator("#example", {
       formatter: "tickCross",
       sorter: "boolean",
       maxWidth: 100,
+      cellEdited: updateFlour,
       formatterParams: {
-        tickElement:`<i class="material-icons green-text">check_box</i>`,
-        crossElement:`<i class="material-icons">check_box_outline_blank</i>`,
+        tickElement: `<i class="material-icons green-text">check_box</i>`,
+        crossElement: `<i class="material-icons">check_box_outline_blank</i>`,
       },
       cellClick: function (e, cell) {
         cell.setValue(!cell.getValue());
+        let col = cell.getColumn();
       },
     },
     {
@@ -77,7 +78,10 @@ var table = new Tabulator("#example", {
       columns: [
         {
           title: "BP (%)",
-          field: "bp",
+          field: "bp0",
+          topCalc: "sum",
+          topCalcParams: { precision: 2 },
+          formatter: formatPercent,
           editor: false,
           editorParams: {
             min: 0,
@@ -86,7 +90,10 @@ var table = new Tabulator("#example", {
         },
         {
           title: "Wt. (g)",
-          field: "wt",
+          field: "wt0",
+          topCalc: "sum",
+          topCalcParams: { precision: 2 },
+          cellEdited: updateFlour,
           editor: "number",
           sorter: "number",
           editorParams: {
@@ -96,41 +103,53 @@ var table = new Tabulator("#example", {
       ],
     },
   ],
-  rowClick: function (e, row) {
-    //trigger an alert message when the row is clicked
-  },
 });
 
+function formatPercent(cell){
+  if (  isNaN((cell.getValue() * 100).toFixed(1))  ){
+    return ""
+  } else
+   return (cell.getValue() * 100).toFixed(1) + "%";
+}
+function loadingCursor(bool) {
+  if (bool) document.body.style.cursor = "wait";
+  else document.body.style.cursor = "default";
+}
 document.addEventListener("DOMContentLoaded", function () {
   var addNewTable = document.getElementById("add-new-table");
   addNewTable.addEventListener("click", addTable);
-  
-  var addRowBtn = document.getElementById('add-row-btn');
-  addRowBtn.addEventListener("click", addRow)
-  var sortBtn = document.getElementById('sort-btn');
-  sortBtn.addEventListener("click", sortRecipe)
-  
-  
+
+  var addRowBtn = document.getElementById("add-row-btn");
+  addRowBtn.addEventListener("click", addRow);
+
+  var sortBtn = document.getElementById("sort-btn");
+  sortBtn.addEventListener("click", sortRecipe);
+
   var elems = document.querySelectorAll(".fixed-action-btn");
   M.FloatingActionButton.init(elems, { direction: "left" });
-  elems = document.querySelectorAll('.collapsible');
+  elems = document.querySelectorAll(".collapsible");
   M.Collapsible.init(elems);
-  init();
+  loadingCursor(false);
 });
-function addRow(){
-    table.addRow({});
-    console.log(table.columnManager,table.columnManager.columns.length)
+function addRow() {
+  updateFlour();
+  table.addRow({});
 }
-function sortRecipe(){
+function sortRecipe() {
+  updateFlour();
   table.setSort(defaultSort);
 }
 
 var fdColumns = {
   title: `Final Dough`,
+  field: "fd",
   columns: [
     {
       title: "BP (%)",
-      field: "bp1",
+      field: "bpf",
+      topCalc: "sum",
+      topCalcParams: { precision: 2 },
+      formatter: formatPercent,
       editor: false,
       editorParams: {
         min: 0,
@@ -139,46 +158,146 @@ var fdColumns = {
     },
     {
       title: "Wt. (g)",
-      field: "wt1",
-      editor: "number",
+      field: "wtf",
+      topCalc: "sum",
+      topCalcParams: { precision: 2 },
+      cellEdited: updateFlour,
+      editor: false,
       sorter: "number",
       editorParams: {
         min: 0,
-      }}
-    ]
-}
+      },
+    },
+  ],
+};
 function addTable(evt) {
-  table.deleteColumn(fdColumns);
-  
-  table.addColumn({
-        title: `<input type="text" value="New Table${++tableLength}" required />`,
-        columns: [
-          {
-            title: "BP (%)",
-            field: "bp1",
-            editor: false,
-            editorParams: {
-              min: 0,
-              max: 100,
-            },
+  loadingCursor(true);
+  if (tableLength !== 0) {
+    table.deleteColumn("bpf");
+    table.deleteColumn("wtf");
+  }
+  tableLength++;
+  table
+    .addColumn({
+      title: ` <div><button class="delete-table-btn btn-flat red-text" id="deleteColumn_${tableLength}">x</button> <input type="text" size="6" value="New Table${tableLength}" required /></div>`,
+      columns: [
+        {
+          title: "BP (%)",
+          field: `bp${tableLength}`,
+          topCalc: "sum",
+          topCalcParams: { precision: 2 },
+          formatter: formatPercent,
+          editor: false,
+          editorParams: {
+            min: 0,
+            max: 100,
           },
-          {
-            title: "Wt. (g)",
-            field: "wt1",
-            editor: "number",
-            sorter: "number",
-            editorParams: {
-              min: 0,
-            },
+        },
+        {
+          title: "Wt. (g)",
+          field: `wt${tableLength}`,
+          topCalc: "sum",
+          topCalcParams: { precision: 2 },
+          cellEdited: updateFlour,
+          editor: "number",
+          sorter: "number",
+          editorParams: {
+            min: 0,
           },
-        ],
-      }).then( function(){
-          table.addColumn(fdColumns);
-    
-      });
+        },
+      ],
+    })
+    .then(function () {
+      table.addColumn(fdColumns);
+      let tableButton = document.getElementById(`deleteColumn_${tableLength}`);
+      tableButton.addEventListener("click", async function () {
+        table.deleteColumn(`bp${tableLength}`);
+        table.deleteColumn(`wt${tableLength}`);
+        tableLength--;
 
-  console.log(evt.target);
+        if (tableLength === 0) {
+          table.deleteColumn(`bpf`);
+          table.deleteColumn(`wtf`);
+        }
+      });
+    });
+  loadingCursor(false);
 }
+
+function updateFlour() {
+  let flourColumn = table.getColumn("isflour");
+  let flourWeight = [];
+  let elseWeight = [];
+
+  let activeRows = [];
+
+  /* Look down isFlour column, traverse and build flour/else totals
+  add any non-zero 
+*/
+
+  flourColumn.getCells().forEach((cell) => {
+    let row = cell.getRow();
+    let data = row.getData();
+    let isFlour = !!cell.getValue();
+    if (row._row.type == "calc") return;
+
+    for (let i = 0; i <= tableLength; i++) {
+      if (isNaN(flourWeight[i])) flourWeight[i] = 0;
+      if (isNaN(elseWeight[i])) elseWeight[i] = 0;
+
+      let thisWeight = Number.parseFloat(data[`wt${i}`]);
+      if (thisWeight > 0) {
+        if (isFlour) {
+          flourWeight[i] += thisWeight;
+          activeRows.push(row);
+        } else {
+          elseWeight[i] += thisWeight;
+          activeRows.push(row);
+        }
+      }
+    }
+    if (tableLength > 0) {
+      let l = tableLength + 1;
+      if (isNaN(flourWeight[l])) flourWeight[l] = 0;
+      if (isNaN(elseWeight[l])) elseWeight[l] = 0;
+
+      let thisWeight = Number.parseFloat(data[`wtf`]);
+      if (thisWeight > 0) {
+        if (isFlour) {
+          flourWeight[l] += thisWeight;
+          activeRows.push(row);
+        } else {
+          elseWeight[l] += thisWeight;
+          activeRows.push(row);
+        }
+      }
+    }
+  });
+console.log(flourWeight, elseWeight);
+/* Calculate Bakers Percentage */
+  activeRows.forEach((row) => {
+    let data = row.getData();
+    let tableTotal = 0;
+    let formulaTotal = 0;
+    for (let i = 0; i <= tableLength; i++) {
+      let bpCell = row.getCell(`bp${i}`);
+      bpCell.setValue(data[`wt${i}`] / flourWeight[i]);
+      if (i> 0 && tableLength > 0){
+        tableTotal +=(!isNaN(data[`wt${i}`]))? Number.parseFloat(data[`wt${i}`]):0;
+      }
+      else if(tableLength > 0)
+        formulaTotal = (!isNaN(data[`wt${i}`]))? Number.parseFloat(data[`wt${i}`]): 0;
+    }
+    console.log(formulaTotal, tableTotal);
+    if(tableLength > 0){
+      row.update({"wtf": formulaTotal - tableTotal})
+      let bpCell = row.getCell(`bpf`);
+      bpCell.setValue(Number.parseFloat(data[`wtf`]) / flourWeight[tableLength + 1]);
+    }
+
+  });
+}
+
 function updateId(value, data, type, params, component) {
   //value - original value of the cell
   //data - the data for the row
@@ -192,30 +311,24 @@ function updateId(value, data, type, params, component) {
   let row = component.getRow();
   if (idx >= 0) {
     row.update({ id: searchTable[idx]._id });
-  }else{
-    row.update({ id: "CREATE" }); 
+  } else {
+    row.update({ id: "CREATE" });
   }
   // data.update({id:searchTable[idx]._id });
   return value; //return the new value for the cell data.
 }
 async function fuzzySearch(term) {
+  if (term.length <= 0)
+    term = "";
+
+  loadingCursor(true);
   try {
     let cInventory = await fetch(
       `/inventory/getFuzzy?search=${term}`
     ).then((res) => res.json());
+    loadingCursor(false);
     return cInventory;
   } catch (err) {
     console.log(err);
   }
 }
-
-async function init() {
-  try {
-      
-    render();
-  } catch (err) {
-    console.log(err);
-  }
-}
-
-function render() {}
